OpenGL rendering stuff. Added OpenGL Shader Program class. Separated GlfwRenderWindow to different file. Double precision for point cloud rendering. Link to GLEW statically. Point cloud rendering (and app pcviewer) definitely not yet finished.

+++ spatiumgl/gfx3d/include/spatiumgl/gfx3d/Renderer.hpp
@@ -35,7 +35,7 @@ namespace spatiumgl {
 				return m_valid;
 			}
 
-			virtual void render(Camera* camera, float aspect) = 0;
+			virtual void render(Camera* camera, double aspect) = 0;
 
 		protected:
 			const RenderObject* m_renderObject;
